changed deleteUser to correctly delete all records for user (used to delete demo user on logout

+++ bin/demoUser/user.js
@@ -5,6 +5,7 @@ const db = require('../../db/models');
 const { Op } = db.Sequelize;
 const { user } = require('../../config');
 const { storyLikes } = require('../../db/seeders/data');
+const { numUsers } = require('../../db/seeders/data/usersData');
 
 async function createUser() {
   // returns a randomly generated user with email/username/hashedPassword
@@ -32,7 +33,13 @@ async function deleteUser(userId) {
   const user = await db.User.findByPk(userId);
   if (!user.demo) throw new Error('Cannot delete non-demo account');
 
-  // get story ids
+  // get array of ids for seeded users
+  const users = [];
+  for (let i = 1; i <= numUsers; i++) {
+    users.push(i);
+  }
+
+  // get story ids of stories by userId
   const storyIds = await db.Story.findAll({
     raw: true,
     attributes: ['id'],
@@ -43,8 +50,12 @@ async function deleteUser(userId) {
     return ids.map(obj => obj.id);
   });
 
-  await db.StoryLike.destroy({ where: { storyId: { [Op.in]: storyIds } } });
-  await db.Comment.destroy({ where: { storyId: { [Op.in]: storyIds } } });
+  // delete likes from all userIds stories OR by userId on any stories
+  await db.StoryLike.destroy({ where: { [Op.or]: [{ storyId: { [Op.in]: storyIds }}, userId] } });
+  // delete comments from all userIds stories OR by userId on any stories
+  await db.Comment.destroy({ where: { [Op.or]: [{ storyId: { [Op.in]: storyIds }}, userId] } });
+
+  // delete all stories owned by userId, follows by userID and finally delete user
   await db.Story.destroy({ where: { id: { [Op.in]: storyIds } } });
   await db.Follow.destroy({ where: { userId } });
   await db.User.destroy({ where: { id: userId } });
